refactor: refactored domain models to be consolidated by domain and wired up stub structure for canary reporting
