+++ packages/client/package.json
@@ -4,7 +4,7 @@
   "description": "The client lib for the Referee UX",
   "author": "Justin Field <fieldju@gmail.com>",
   "homepage": "https://example.com/dashboard/",
-  "proxy": "http://localhost:3000",
+  "proxy": "http://localhost:3001",
   "license": "UNLICENSED",
   "private": true,
   "dependencies": {
