+++ packages/client/src/components/App.tsx
@@ -11,12 +11,13 @@ import Landing from './Landing';
 import ModalViewer from './ModalViewer';
 import ConfigEditor from './config/ConfigEditor';
 import CanaryExecutor from './canary-executor/CanaryExecutor';
-import CanaryExecutorResults from './results/CanaryExecutorResults';
+import CanaryExecutorResults from './canary-executor/CanaryExecutorResults';
 import Docs from './docs/Docs';
 
 import './App.scss';
 import 'typeface-assistant/index.css';
 import 'bootstrap/dist/css/bootstrap.css';
+import CanaryReportViewer from './reports/canary/CanaryReportViewer';
 
 interface AppProps {}
 
@@ -32,6 +33,7 @@ export default class App extends React.Component<AppProps> {
               <div className="content-container-inner-wrapper">
                 <Route exact={true} path="/" component={Landing} />
                 <Route path="/docs/:path(.*)" component={Docs} />
+                <Route path="/reports/canary/:executionId(.*)" component={CanaryReportViewer} />
                 <Route path="/config/edit" component={ConfigEditor} />
                 <Route path="/dev-tools/canary-executor/setup" component={CanaryExecutor} />
                 <Route
