--- packages/client/src/domain/CanaryExecutionRequestTypes.ts
@@ -1,23 +1,0 @@
-import { KvMap } from './CustomTypes';
-
-export interface CanaryExecutionRequest {
-  scopes: {
-    default: {
-      controlScope: CanaryScope;
-      experimentScope: CanaryScope;
-    };
-  };
-  thresholds: {
-    marginal: number;
-    pass: number;
-  };
-}
-
-export interface CanaryScope {
-  scope: string;
-  location: string;
-  step: number;
-  start: string;
-  end: string;
-  extendedScopeParams: KvMap<string>;
-}
