--- packages/client/src/domain/CanaryExecutionResponse.ts
@@ -1,3 +1,0 @@
-export interface CanaryExecutionResponse {
-  canaryExecutionId: string;
-}
