--- packages/client/src/domain/KayentaCredential.ts
@@ -1,9 +1,0 @@
-export default interface KayentaCredential {
-  name: string;
-  supportedTypes: string[];
-  objects: any;
-  metadata: any;
-  type: string;
-  recommendedLocations: string[];
-  locations: string[];
-}
