+++ packages/client/src/services/FetchCanaryResultsService.ts
@@ -1,32 +1,32 @@
 import { stores } from '../stores';
 import { kayentaApiService } from './index';
 import { delay } from 'q';
-import { CanaryExecutionStatusResponse } from '../domain/CanaryExecutionStatusResponse';
+import { CanaryExecutionStatusResponse } from '../domain/Kayenta';
 
-const { configEditorStore, canaryExecutorStore, resultsStore } = stores;
+const { configEditorStore, canaryExecutorStore } = stores;
 let response: CanaryExecutionStatusResponse | any = {};
 const SUCCESS = 'succeeded';
 
 export default class FetchCanaryResultsService {
-  async fetchCanaryResults(canaryExecutionId: string): Promise<void> {
+  async pollForCanaryExecutionComplete(canaryExecutionId: string): Promise<void> {
     kayentaApiService.fetchCredentials().then(data => canaryExecutorStore.setKayentaCredentials(data));
-    resultsStore.setCanaryExecutionId(canaryExecutionId);
+    canaryExecutorStore.setCanaryExecutionId(canaryExecutionId);
 
     do {
       const data = async () => {
         await delay(1000);
-        response = await kayentaApiService.fetchCanaryRunStatusAndResults(resultsStore.canaryExecutionId);
-        resultsStore.updateStageStatus(response.stageStatus);
+        response = await kayentaApiService.fetchCanaryExecutionStatusResponse(canaryExecutorStore.canaryExecutionId);
+        canaryExecutorStore.updateStageStatus(response.stageStatus);
       };
       await data();
     } while (!response.complete);
 
     if (response.status === SUCCESS) {
-      resultsStore.updateCanaryExecutionStatusResponse(response);
+      canaryExecutorStore.updateCanaryExecutionStatusResponse(response);
       configEditorStore.setCanaryConfigObject(response.config);
       canaryExecutorStore.setCanaryExecutionRequestObject(response.canaryExecutionRequest);
     }
 
-    resultsStore.updateResultsRequestComplete();
+    canaryExecutorStore.updateResultsRequestComplete();
   }
 }
