+++ packages/client/src/types/hf.d.ts
@@ -1,0 +1,3 @@
+declare module 'human-format' {
+  export default function humanFormat(input: number): string;
+}
